+++ src/data/database.ts
@@ -1146,6 +1146,10 @@ export type Database = {
         Args: { _import_note: string; _submissions: Json; _venueid: string }
         Returns: Json
       }
+      can_approve_assignment: {
+        Args: { _role: string; _submission: string }
+        Returns: boolean
+      }
       complete_assignment: {
         Args: {
           _assignment_id: string
@@ -1188,7 +1192,6 @@ export type Database = {
       export_scholar_data: { Args: { _scholar?: string }; Returns: Json }
       forget_scholar: { Args: { _scholar: string }; Returns: Json }
       isadmin: { Args: { _venueid: string }; Returns: boolean }
-      isapprover: { Args: { _roleid: string }; Returns: boolean }
       isassigned: { Args: { _submissionid: string }; Returns: boolean }
       isauthor: { Args: { _submissionid: string }; Returns: boolean }
       isconflicted: { Args: { _submissionid: string }; Returns: boolean }
@@ -1198,6 +1201,7 @@ export type Database = {
         Returns: boolean
       }
       ispriorityzero: { Args: { _venueid: string }; Returns: boolean }
+      isroleapprovervolunteer: { Args: { _roleid: string }; Returns: boolean }
       issteward: { Args: never; Returns: boolean }
       mark_submission_done: {
         Args: {